+++ libs/ui-inputs/src/index.ts
@@ -1,7 +1,5 @@
 export * from './lib/tag-input-box';
-// export * from './lib/info-box';
-export * from './lib/ui-inputs';
-export * from './lib/input-box/input-box';
+export * from './lib/input-box';
 export * from './lib/switch';
 export * from './lib/drop-down';
 export * from './lib/types';
