--- libs/ui-inputs/src/lib/input-box/input-box.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import InputBox from './input-box';
-
-describe('InputBox', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<InputBox />);
-    expect(baseElement).toBeTruthy();
-  });
-});
