--- libs/ui-inputs/src/lib/ui-inputs.tsx
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-/* eslint-disable-next-line */
-export interface UiInputsProps {}
-
-const StyledUiInputs = styled.div``;
-
-export function UiInputs(props: UiInputsProps) {
-  return (
-    <StyledUiInputs>
-      <h1>Welcome to UiInputs!</h1>
-    </StyledUiInputs>
-  );
-}
-
-export default UiInputs;
